requirements.txt
aiogram>=3.3.0
pydantic>=2.8.0,<3.0.0
pydantic-core>=2.23.0
ccxt>=4.0.0
pandas>=2.0.0
pandas-ta>=0.4.67b0
numpy>=1.24.0
python-dotenv>=1.0.0
cryptography>=41.0.0
aiofiles>=23.0.0
aiohttp>=3.9.0
matplotlib>=3.7.0
mplfinance>=0.12.10b0
